Web.Slack

Description

This library contains functionality for writing slack bots which respond to Events sent by the RTM API.

A basic example of how to use this library is presented below. Other examples can be found in the "examples" directory.

Confusingly, Slack exposes a number of APIs which provide different (but overlapping) functionality. This library is mostly about the RTD API, but it also has some very limited support for using the Web API.

The Slack RTD API

The Slack Real-Time Data API. Once the connection is established, the server pushes notifications to the client over a websocket. You can recieve these notifications by calling getNextEvent.

The RTD API also allows the client to send some messages back over the websocket connection. At the moment, it only really allows the sending of simple messages. For more complex interactions, you still have to use the Web API (see below).

getNextEvent :: MonadSlack m => m Event

Returns the next event in the queue. If none is available, blocks until a new event is recieved.

getSession :: MonadSlack m => m SlackSession

When the connection is established, the slack server sends a bunch of session information. This is accessible here.

(Caveat: this information represents the state as of the beginning of the session; it is liable to become stale. If you care about keeping an up-to-date view of this stuff, you need to track changes to it using getNextEvent.)

sendMessage :: MonadSlack m => ChannelId -> Text -> m ()

Send a simple slack message.

sendPing :: MonadSlack m => m ()

Send a ping to the server, which should respond by sending a Pong event.

data SlackSession

Constructors

SlackSession

Fields

_slackSelf :: Self

_slackTeam :: Team

_slackUsers :: [User]

_slackLatestEventTs :: SlackTimeStamp

_slackChannels :: [Channel]

_slackGroups :: [Channel]

_slackIms :: [IM]

_slackBots :: [Bot]

_slackCacheVersion :: Text

data Event where

Constructors

Hello :: Event

Message :: ChannelId -> Submitter -> Text -> SlackTimeStamp -> Maybe Subtype -> Maybe Edited -> Event

HiddenMessage :: ChannelId -> Submitter -> SlackTimeStamp -> Maybe Subtype -> Event

ChannelMarked :: ChannelId -> SlackTimeStamp -> Event

ChannelCreated :: Channel -> Event

ChannelJoined :: Channel -> Event

ChannelLeft :: ChannelId -> Event

ChannelDeleted :: ChannelId -> Event

ChannelRename :: ChannelRenameInfo -> Event

ChannelArchive :: ChannelId -> UserId -> Event

ChannelUnarchive :: ChannelId -> UserId -> Event

ChannelHistoryChanged :: SlackTimeStamp -> SlackTimeStamp -> SlackTimeStamp -> Event

ImCreated :: UserId -> IM -> Event

ImOpen :: UserId -> IMId -> Event

ImClose :: UserId -> IMId -> Event

ImMarked :: IMId -> SlackTimeStamp -> Event

ImHistoryChanged :: SlackTimeStamp -> SlackTimeStamp -> SlackTimeStamp -> Event

GroupJoined :: Channel -> Event

GroupLeft :: Channel -> Event

GroupOpen :: UserId -> ChannelId -> Event

GroupClose :: UserId -> ChannelId -> Event

GroupArchive :: ChannelId -> Event

GroupUnarchive :: ChannelId -> Event

GroupRename :: ChannelRenameInfo -> Event

GroupMarked :: ChannelId -> SlackTimeStamp -> Event

GroupHistoryChanged :: SlackTimeStamp -> SlackTimeStamp -> SlackTimeStamp -> Event

FileCreated :: File -> Event

FileShared :: FileReference -> Event

FileUnshared :: File -> Event

FilePublic :: FileReference -> Event

FilePrivate :: FileId -> Event

FileChange :: File -> Event

FileDeleted :: FileId -> SlackTimeStamp -> Event

FileCommentAdded :: File -> Comment -> Event

FileCommentEdited :: File -> Comment -> Event

FileCommentDeleted :: File -> CommentId -> Event

PresenceChange :: UserId -> Presence -> Event

ManualPresenceChange :: Presence -> Event

PrefChange :: Pref -> Event

UserChange :: User -> Event

TeamJoin :: User -> Event

ReactionAdded :: UserId -> Text -> UserId -> Item -> SlackTimeStamp -> Event

ReactionRemoved :: UserId -> Maybe Text -> Item -> SlackTimeStamp -> Event

StarAdded :: UserId -> Item -> SlackTimeStamp -> Event

StarRemoved :: UserId -> Item -> SlackTimeStamp -> Event

EmojiChanged :: SlackTimeStamp -> Event

CommandsChanged :: SlackTimeStamp -> Event

TeamPrefChange :: Pref -> Event

TeamRenameEvent :: Text -> Event

TeamDomainChange :: URL -> Domain -> Event

EmailDomainChange :: Domain -> SlackTimeStamp -> Event

BotChanged :: Bot -> Event

BotAdded :: Bot -> Event

AccountsChanged :: Event

UserTyping :: ChannelId -> UserId -> Event

MessageResponse :: Int -> SlackTimeStamp -> Text -> Event

MessageError :: Int -> SlackError -> Event

StatusChange :: UserId -> Text -> SlackTimeStamp -> Event

Pong :: Time -> Event

ReconnectUrl :: URL -> Event

TeamMigrationStarted :: Event

PinAdded :: Event

PinRemoved :: Event

NoEvent :: Event

UnknownEvent :: Value -> Event

module Web.Slack.Types

The Slack Web API

The Slack Web API is an HTTP REST API. It supports more user operations than the RTD API. This API is still not very well supported by this library.

chat_postMessage :: (MonadError String m, MonadIO m) => SlackConfig -> ChannelId -> Text -> [Attachment] -> m (Response ByteString)

module Web.Slack.WebAPI

Running your code

The API exposed by this library is fairly generic, and is designed to support writing code in two different styles: by using a Slack monad, or by passing around a SlackHandle.

class (MonadReader SlackHandle m, MonadIO m) => MonadSlack m

A monad which supports performing slack-api operations. Some useful instances include:

Slack a

SlackHandle -> IO a

RWST SlackHandle () MyState IO a

getConfig :: MonadSlack m => m SlackConfig

Retrieve the config used to initiate the session.

data SlackConfig

Configuration options needed to connect to the Slack API

Constructors

SlackConfig

Fields

_slackApiToken :: String

API Token for Bot

Monadic style

type Slack = ReaderT SlackHandle IO

runSlack :: SlackConfig -> Slack a -> IO a

The following example is for a bot which echoes every message it sees back to the channel it case from.

main :: IO ()
main = runSlack myConfig echoBot

myConfig :: SlackConfig
myConfig = SlackConfig { _slackApiToken = "your API token here" }

echoBot :: Slack ()
echoBot = do
    event <- getNextEvent
    case event of
        (Message cid _ msg _ _ _) -> sendMessage cid msg
        _ -> return ()
    echoBot

Handle-based style

data SlackHandle

withSlackHandle :: SlackConfig -> (SlackHandle -> IO a) -> IO a

This is the same example as in the previous section, but written using a handle-based style. This approach is more verbose, but can be more flexible (eg. consider forking threads which share a handle). The API exposed by this module can be used with either style.

main :: IO ()
main = withSlackHandle myConfig echoBot

myConfig :: SlackConfig
myConfig = SlackConfig { _slackApiToken = "your API token here" }

echoBot :: SlackHandle -> IO ()
echoBot h = do
    event <- getNextEvent h
    case event of
        (Message cid _ msg _ _ _) -> sendMessage cid msg h
        _ -> return ()
    echoBot h