Web.Slack.Internal

Documentation

data SlackHandle

withSlackHandle :: SlackConfig -> (SlackHandle -> IO a) -> IO a

config :: SlackHandle -> SlackConfig

session :: SlackHandle -> SlackSession

nextEvent :: SlackHandle -> IO Event

Returns the next event. If the queue is empty, blocks until an event is recieved.

sendMessage :: SlackHandle -> ChannelId -> Text -> IO ()

Send a message to the specified channel.

If the message is longer than 4000 bytes then the connection will be closed.

sendPing :: SlackHandle -> IO ()

Send a ping packet to the server The server will respond with a pong Event.