Web.Slack.Monad

Description

The API exposed in this module is a generalisation of the one in Web.Slack.Internal.

Documentation

type Slack = ReaderT SlackHandle IO

runSlack :: SlackConfig -> Slack a -> IO a

class (MonadReader SlackHandle m, MonadIO m) => MonadSlack m

A monad which supports performing slack-api operations. Some useful instances include:

Slack a

SlackHandle -> IO a

RWST SlackHandle () MyState IO a

getConfig :: MonadSlack m => m SlackConfig

Retrieve the config used to initiate the session.

getSession :: MonadSlack m => m SlackSession

When the connection is established, the slack server sends a bunch of session information. This is accessible here.

(Caveat: this information represents the state as of the beginning of the session; it is liable to become stale. If you care about keeping an up-to-date view of this stuff, you need to track changes to it using getNextEvent.)

getNextEvent :: MonadSlack m => m Event

Returns the next event in the queue. If none is available, blocks until a new event is recieved.

sendMessage :: MonadSlack m => ChannelId -> Text -> m ()

Send a simple slack message.

sendPing :: MonadSlack m => m ()

Send a ping to the server, which should respond by sending a Pong event.